CLOUDP-268809: IP Access List Controller (#2063)

+++ internal/controller/atlasproject/ipaccess_list.go
@@ -1,12 +1,15 @@
 package atlasproject
 
 import (
+	"encoding/json"
 	"errors"
+	"fmt"
 	"reflect"
 
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/api"
 	akov2 "github.com/mongodb/mongodb-atlas-kubernetes/v2/api/v1"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/api/v1/status"
+	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/customresource"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/workflow"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/translation/ipaccesslist"
 )
@@ -15,9 +18,10 @@ const ipAccessStatusPending = "PENDING"
 const ipAccessStatusFailed = "FAILED"
 
 type ipAccessListController struct {
-	ctx     *workflow.Context
-	project *akov2.AtlasProject
-	service ipaccesslist.IPAccessListService
+	ctx         *workflow.Context
+	project     *akov2.AtlasProject
+	service     ipaccesslist.IPAccessListService
+	lastApplied ipaccesslist.IPAccessEntries
 }
 
 // reconcile dispatch state transitions
@@ -52,6 +56,10 @@ func (i *ipAccessListController) configure(current, desired ipaccesslist.IPAcces
 	}
 
 	for key, entry := range current {
+		if _, ok := i.lastApplied[key]; !ok {
+			continue
+		}
+
 		if _, ok := desired[key]; !ok {
 			err = i.service.Delete(i.ctx.Context, i.project.ID(), entry)
 			if err != nil {
@@ -123,11 +131,61 @@ func handleIPAccessList(ctx *workflow.Context, project *akov2.AtlasProject) work
 	ctx.Log.Debug("starting ip access list processing")
 	defer ctx.Log.Debug("finished ip access list processing")
 
+	skipped, err := shouldIPAccessListSkipReconciliation(project)
+	if err != nil {
+		return workflow.Terminate(workflow.Internal, err)
+	}
+
+	if skipped {
+		ctx.EnsureStatusOption(status.AtlasProjectExpiredIPAccessOption(nil))
+		ctx.UnsetCondition(api.IPAccessListReadyType)
+
+		return workflow.OK()
+	}
+
+	lastApplied, err := mapLastAppliedIPAccessList(project)
+	if err != nil {
+		return workflow.Terminate(workflow.Internal, fmt.Errorf("failed to get last applied configuration: %w", err))
+	}
+
 	c := ipAccessListController{
-		ctx:     ctx,
-		project: project,
-		service: ipaccesslist.NewIPAccessList(ctx.SdkClient.ProjectIPAccessListApi),
+		ctx:         ctx,
+		project:     project,
+		service:     ipaccesslist.NewIPAccessList(ctx.SdkClient.ProjectIPAccessListApi),
+		lastApplied: lastApplied,
 	}
 
 	return c.reconcile()
+}
+
+func shouldIPAccessListSkipReconciliation(atlasProject *akov2.AtlasProject) (bool, error) {
+	lastSkippedSpec := akov2.AtlasProjectSpec{}
+	lastSkippedSpecString, ok := atlasProject.Annotations[customresource.AnnotationLastSkippedConfiguration]
+	if ok {
+		if err := json.Unmarshal([]byte(lastSkippedSpecString), &lastSkippedSpec); err != nil {
+			return false, fmt.Errorf("failed to parse last skipped configuration: %w", err)
+		}
+
+		return len(lastSkippedSpec.ProjectIPAccessList) == 0, nil
+	}
+
+	return false, nil
+}
+
+func mapLastAppliedIPAccessList(atlasProject *akov2.AtlasProject) (ipaccesslist.IPAccessEntries, error) {
+	lastApplied, err := lastSpecFrom(atlasProject, customresource.AnnotationLastAppliedConfiguration)
+	if err != nil {
+		return nil, err
+	}
+
+	if lastApplied == nil || len(lastApplied.ProjectIPAccessList) == 0 {
+		return nil, nil
+	}
+
+	entries, err := ipaccesslist.NewIPAccessEntries(lastApplied.ProjectIPAccessList)
+	if err != nil {
+		return nil, err
+	}
+
+	return entries, nil
 }
